translucent -> true
