+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,6 @@
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
             android:theme="@style/Theme.Hashin">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
