+++ app/src/main/java/com/karan/hashin/MainActivity.kt
@@ -35,24 +35,24 @@ class MainActivity : ComponentActivity() {
                 val authViewModel: AuthViewModel = viewModel()
                 val isAuthenticated = authViewModel.isAuthenticated.collectAsState().value
 
-                Scaffold(
-                    topBar = {
-                        Log.d("#ined", "isAuthenticated ? $isAuthenticated")
-                        if (isAuthenticated) {
-                            TopAppBar(Modifier)
-                        }
-                    },
-                    bottomBar = {
-                        Log.d("#ined", "isAuthenticated ? $isAuthenticated")
-                        if (isAuthenticated) {
-                            BottomAppBar(Modifier)
-                        }
-                    },
-                    modifier = Modifier
-                        .padding(WindowInsets.systemBars.asPaddingValues())
-                ) { pd ->
-                    NavGraph(pd, navController)
-                }
+//                Scaffold(
+//                    topBar = {
+//                        Log.d("#ined", "isAuthenticated ? $isAuthenticated")
+//                        if (isAuthenticated) {
+//                            TopAppBar(Modifier)
+//                        }
+//                    },
+//                    bottomBar = {
+//                        Log.d("#ined", "isAuthenticated ? $isAuthenticated")
+//                        if (isAuthenticated) {
+//                            BottomAppBar(Modifier)
+//                        }
+//                    },
+//                    modifier = Modifier
+//                        .padding(WindowInsets.systemBars.asPaddingValues())
+//                ) { pd ->
+                    NavGraph(navController)
+//                }
             }
         }
     }
