+++ app/src/main/java/com/karan/hashin/screens/Home.kt
@@ -9,6 +9,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -16,33 +17,27 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.LineHeightStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
+import com.karan.hashin.components.BottomAppBar
 import com.karan.hashin.components.Element
+import com.karan.hashin.components.TopAppBar
 import com.karan.hashin.ui.theme.HashinTheme
 
 
 @Composable
 fun HomeScreen(modifier: Modifier = Modifier) {
 
+    Vault(Modifier)
 
-    Box(
-        contentAlignment = Alignment.Center,
-        modifier = Modifier
-            .fillMaxSize()
-            .navigationBarsPadding()
-    ) {
-
-        LazyColumn(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            modifier = Modifier
-                .fillMaxWidth()
-        ) {
-            val data = listOf(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12)
-
-            items(count = data.size, key = { it }) {
-                Element()
-            }
-        }
-    }
+//    Scaffold(
+//        topBar = {
+//            TopAppBar(Modifier)
+//        },
+//        bottomBar = {
+//            BottomAppBar(Modifier)
+//        }
+//    ) { pd ->
+//
+//    }
 }
 
 @Preview(showBackground = true, showSystemUi = true)
