+++ app/src/main/java/com/karan/hashin/screens/Vault.kt
@@ -1,0 +1,34 @@
+package com.karan.hashin.screens
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.navigationBarsPadding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import com.karan.hashin.components.Element
+
+@Composable
+fun Vault(modifier: Modifier = Modifier) {
+    Box(
+        contentAlignment = Alignment.Center,
+        modifier = Modifier
+            .fillMaxSize()
+            .navigationBarsPadding()
+    ) {
+
+        LazyColumn(
+            horizontalAlignment = Alignment.CenterHorizontally,
+            modifier = Modifier
+                .fillMaxWidth()
+        ) {
+            val data = listOf(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12)
+
+            items(count = data.size, key = { it }) {
+                Element()
+            }
+        }
+    }
+}
